+++ Binaries/dcmtk-source/dcmdata/dcbytstr.h
@@ -108,7 +108,7 @@ class DcmByteString
     /** get value multiplicity
      *  @return number of string components (separated by a backslash)
      */
-    virtual unsigned long getVM();
+    virtual unsigned int getVM();
 
     /** get length of the stored value.
      *  Trailing spaces (padding characters) are ignored for the "real" length.
@@ -166,7 +166,7 @@ class DcmByteString
      *  @return status, EC_Normal if successful, an error code otherwise
      */
     virtual OFCondition getOFString(OFString &stringVal,
-                                    const unsigned long pos,
+                                    const unsigned int pos,
                                     OFBool normalize = OFTrue);
 
     /** get a pointer to the current string value.
@@ -274,7 +274,7 @@ const OFBool MULTIPART = OFTrue;
  */
 OFCondition getStringPart(OFString &result,
                           const char *orgStr,
-                          const unsigned long pos);
+                          const unsigned int pos);
 
 
 /** normalize the given string value, i.e. remove leading and/or trailing spaces
@@ -404,9 +404,9 @@ void normalizeString(OFString &string,
 **   overloaded get methods in all derived classes of DcmElement.
 **   So the interface of all value representation classes in the
 **   library are changed rapidly, e.g.
-**   OFCondition get(Uint16 & value, const unsigned long pos);
+**   OFCondition get(Uint16 & value, const unsigned int pos);
 **   becomes
-**   OFCondition getUint16(Uint16 & value, const unsigned long pos);
+**   OFCondition getUint16(Uint16 & value, const unsigned int pos);
 **   All (retired) "returntype get(...)" methods are deleted.
 **   For more information see dcmdata/include/dcelem.h
 **
