+++ Binaries/dcmtk-source/dcmdata/dclist.cc
@@ -264,11 +264,11 @@ DcmObject *DcmList::seek( E_ListPos pos )
 // ********************************
 
 
-DcmObject *DcmList::seek_to(unsigned long absolute_position)
+DcmObject *DcmList::seek_to(unsigned int absolute_position)
 {
-    const unsigned long tmppos = absolute_position < cardinality ? absolute_position : cardinality;
+    const unsigned int tmppos = absolute_position < cardinality ? absolute_position : cardinality;
     seek( ELP_first );
-    for (unsigned long i = 0; i < tmppos; i++)
+    for (unsigned int i = 0; i < tmppos; i++)
         seek( ELP_next );
     return get( ELP_atpos );
 }
